✨ update : ProjectData를 '@/types/project'에서 가져오도록 수정 및 관련 타입 정의 추가 #3
